UI: Refactor (module structure)
