+++ HoloRepositoryUI/src/components/app/new_hologram/shared/DetailsDeclarationStep.tsx
@@ -1,8 +1,8 @@
 import React, { Component } from "react";
-import { IPatient } from "../../../types";
+import { IPatient } from "../../../../types";
 
-import samplePatients from "../../../__tests__/samples/samplePatients.json";
-import samplePatientsWithHolograms from "../../../__tests__/samples/samplePatientsWithHolograms.json";
+import samplePatients from "../../../../__tests__/samples/samplePatients.json";
+import samplePatientsWithHolograms from "../../../../__tests__/samples/samplePatientsWithHolograms.json";
 import {
   DatePicker,
   DayOfWeek,
