+++ HoloRepositoryUI/src/components/app/new_hologram/shared/NewHologramControlsAndProgress.tsx
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import { DefaultButton, PrimaryButton } from "office-ui-fabric-react";
 import { Steps } from "antd";
-import { IHologramCreationStep } from "./NewHologramPage";
+import { IHologramCreationStep } from "../NewHologramPage";
 
 const { Step } = Steps;
